Make transfer encoding configurable
With this change we enable users to configure the transfer encoding to
use. So far this encoding has been set automatically by the lettre
crate, which supposedly picks the more efficient encoding. However, it
can only do so much, and detection of whether something is just plain
8bit or actually base64 encoded is not being done.
As such, we allow users to set the transfer encoding to use via the
configuration file. The setting is not exposed via the command line,
mostly because it mostly depends on the "filter" configured what the
right transfer encoding is, and filters are only configurable via the
configuration file.