Remove line transformation hack
Our string replacement converting line endings from one format to
another was mostly a hack. It required conversion from a byte array into
a string and it was not strictly speaking correct, because it was not
idempotent. It also turned out that setting the correct transfer
encoding, which we made possible earlier, is sufficient to fix any
issues we have seen earlier with respect to line endings.
To that end, this change removes the line ending transformation kludge.