Reword comment regarding potential errors
The comment we have on the path sending an email with all errors
encountered so far is not worded super precisely: there is no guarantee
of an error, as seems to be implied. Rather, there is the potential for
one.
With this change we reword the comment slightly to make it a bit less
confusing.