Refactor email sending internals
Moving forward we would like to introduce a proper split between the
command line interface and the underlying functionality that is also
provided in the form of a library. To get to this state, this change
refactors the email sending internals in such a way, to make extraction
of the CLI functionality a bit easier.