Factor out a proper crate for command line interface
This change factors out the command line functionality into a separate
crate. This crate inherits the original name, mail-message, while the
library will use maily moving forward.
As part of this rework we establish a first version of some kind of
public API to the underlying library. This library interface is expected
to be refined more in the future.