Make content type configurable
We expect that most of the time we only want to send plain text messages
from the command line. However, that may not always be the case. If
something else is being sent, it makes sense to allow the user to
specify the content type to use for the email.
To that end, this change makes this property configurable via program
arguments.