cli: Only print "enter message" output in terminal
We explicitly support reading data from stdin if no message is provided
on the command line. To make it clear that input is accepted, we
currently print a message. However, in case data is just piped from
another program, such a message serves no purpose and can actively cause
confusion.
With this change we only print it if stdin is connected to an actual
terminal.