Make lettre normalize line endings
For a long time we have struggled with emails being rejected due to
incorrect line endings. After a bit more research it came to light that
the lettre crate can actually normalize line endings correctly, but does
so only when the input is a string (as opposed to a vector of bytes).
To make use of this functionality, make sure to interpret the message to
send as string (if possible) and feed it to lettre in that form.