Add PGP encryption support
This change adds support for encrypting email messages using PGP before
sending them. We originally intended for this to be done without direct
involvement or knowledge of the library, but that seems less and less
feasible. At this point encryption requires an explicit keybox path to
be passed in, so that the data necessary for encryption can be looked up
there. We do not currently support looking up global GnuPG configuration
or similar.