Remove support for specifying transfer encoding
With the switch towards letting lettre handle line ending conversions,
the belief is that we no longer need to provide the means for specifying
the content transfer encoding. Said encoding should ideally be an
implementation detail, and it may differ depending on various internal
conditions, rendering a single option for controlling it questionable.
This change removes the logic for giving the user control over what
encoding to use in favor of a fully automatic detection.