Store Cow objects in Account type
So far our Account type was solely a means for grouping certain input
data, in which case it made sense for it to only contain references.
With an upcoming change objects of this type will also be deserialized,
in which case we need them to hold owned objects.
With this change we convert the various reference members it contains
into Cow objects to enable both use cases.