Introduce deserializable configuration
So far it was only the CLI that had a notion of a persistent
configuration. However, what we ultimately want is for the library to be
zero-config by default and to just pick up whatever is configured on the
system.
To that end, this change adds and exports a Config type that determines
the configuration file structure. It furthermore defines the path to a
system-wide configuration that can be loaded on-demand and destructed
into the corresponding inputs to the send_email function. All this
functionality is provided behind the newly introduced 'config' feature
switch.