cli: Re-use maily's config support
Now that the maily crate comes with its own configuration
infrastructure, we can adjust the mail-message CLI to make use of it.
This change does precisely that, eliminating a bunch of duplication in
the process. While at it, we also get rid of the support for
user-specific configuration. For now only the system-wide configuration
(or whatever is provided as a program argument) is honored.